fix: base publication flow collection on piece, not agendaitem
If we go via the agendaitem (as before) the link between agenda and
publication flow goes through a case, which is a lot more likely to
exist in cases where we don't want to propagate.

E.g. a case with a (decisionmakinFlow with a) subcase that is already
part of a propagated agenda. If a new subcase is created and submitted
on a new agenda that is not released yet, any new pub flow on this new
agendaitem will immediately be propagated even though other profiles may
not even see the agenda yet (because there already exists a link via the
case).